requirements.txt
lmnt
openai-whisper

